+++ app.json
@@ -10,7 +10,7 @@
         "frame": true,
         "defaultHeight": 600,
         "defaultWidth": 800,
-        "preload": "http://localhost:8080/ssf.js"
+        "preload": "http://localhost:8080/bundle/bundle.js"
     },
     "runtime": {
         "arguments": "--noerrdialogs",
