+++ js/notify.js
@@ -1,27 +1,8 @@
-/* override window.open to fix name issue */
-var originalOpen = window.open;
-window.open = (...args) => {
-   let w = originalOpen.apply(this, args);
-    //Try catch for cross domain safeguard
-    try {
-       w.name = args[1];
-     } catch (e) {
-        console.log(e)
-     }
-
-     return w;
-}
-
-/*
-  to do:
-    - events for notifications
-    - fix pop-out (talk to Lynn)
-
-*/
 /*
 * Class representing a Symphony notification
 */
-class Notify {
+
+module.exports = class Notify {
 
     constructor(title,options){
         console.log("SSF Notify " + JSON.stringify(title) + JSON.stringify(options));
@@ -81,59 +62,3 @@ class Notify {
         // How is this different from close?
     }
 }
-
-/*
-  core symphony API
-*/
-window.SYM_API = {
-    Notification:Notify,
-
-    setBadgeCount:function(Number) {
-        console.log("SSF Badgecount " + Number);
-    },
-    // TESTED AND SHOWS UP
-    ScreenSnippet:function(cb) {
-        console.log("SSF Screen Snippet requested");
-        fin.desktop.Window.getCurrent().getSnapshot(base64Snapshot => {
-            console.log('Window Snapshot Taken');
-            if(cb) {
-                cb(base64Snapshot);
-            }
-        })
-    },
-    activate:function() {
-        console.log("SSF Activate!");
-        fin.desktop.Window.getCurrent().bringToFront();
-    },
-    //undoced
-    registerLogger:function() {
-        console.log("SSF registerLogger!!");
-    },
-
-    // TESTED AND WORKS
-    registerBoundsChange:function(callback) {
-        console.log("SSF boundschange!")
-        var cb = callback;
-        fin.desktop.Window.getCurrent().addEventListener("bounds-changed", obj => {
-        cb({x:obj.left,
-            y:obj.top,
-            width:obj.width,
-            height:obj.height,
-            windowName:obj.name});
-        })
-    },
-
-    getVersionInfo: function() {
-        return new Promise((resolve, reject) => {
-            let version = {
-                containerIdentifier: "SymphonyOpenFin",
-                containerVer: "0.0.1",
-                apiVer: "1.0.0"
-            }
-            resolve(version)
-        })
-    }
-}
-
-window.ssf = window.SYM_API;
-window.ssf.activate();
