+++ js/window.js
@@ -1,0 +1,13 @@
+/* override window.open to fix name issue */
+var originalOpen = window.open;
+module.exports = (...args) => {
+   let w = originalOpen.apply(this, args);
+    //Try catch for cross domain safeguard
+    try {
+       w.name = args[1];
+     } catch (e) {
+        console.log(e)
+     }
+
+     return w;
+}
